requirements.txt
streamlit
streamlit-webrtc
opencv-python
numpy
cvzone
tensorflow
keras
pyttsx3
pyenchant